got the card working now needs css
