--- vite-pokeapi/src/components/HomeCard.tsx
@@ -1,7 +1,0 @@
-const HomeCard = () =>{
-    return (
-        <h1>this is the card component</h1>
-    )
-}
-
-export default HomeCard
